// apps/web/src/lib/api/generated/model/role.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Inventory & Warehouse System API
 * REST API for the Inventory & Warehouse System
 * OpenAPI spec version: 1.0
 */

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  WAREHOUSE_MANAGER: 'WAREHOUSE_MANAGER',
  STAFF: 'STAFF',
  SUPPLIER: 'SUPPLIER',
} as const;